+++ R/zzz.R
@@ -14,27 +14,13 @@ To proceed, install Nix or use a system where it is available."
     )
   }
 
-  nix_conf_path <- "~/.config/nix/nix.conf"
-
-  if (
-    nix_conf_exists(nix_conf_path) &&
-      !is_cachix_configured(readLines(nix_conf_path))
-  ) {
+  # Only show cachix message if Nix is installed but cache is not configured
+  if (nix_build_installed() && !is_cachix_configured_anywhere()) {
     packageStartupMessage(
       "To speed up builds, configure the rstats-on-nix cache by running:
   rix::setup_cachix()
-See vignette(\"b1-setting-up-and-using-rix-on-linux-and-windows\") or
-vignette(\"b2-setting-up-and-using-rix-on-macos\") for full instructions.
-(You only need to do this once per machine.)"
-    )
-  }
-
-  if (!nix_conf_exists(nix_conf_path) && nix_build_installed()) {
-    packageStartupMessage(
-      "To speed up builds, configure the rstats-on-nix cache by running:
-  rix::setup_cachix()
-See vignette(\"b1-setting-up-and-using-rix-on-linux-and-windows\") or
-vignette(\"b2-setting-up-and-using-rix-on-macos\") for full instructions.
+See vignette(\"setting-up-linux-windows\") or
+vignette(\"setting-up-macos\") for full instructions.
 (You only need to do this once per machine.)"
     )
   }